mode for stubbing all API calls
This obviates the need to stand up the server to run tests. To fake api
calls, adjust your Config/<environment>/Settings.plist
