+++ src/BikeTagTests/Models/SpotTest.swift
@@ -2,45 +2,6 @@ import XCTest
 import UIKit
 import CoreLocation
 
-class FakeSpotsService: SpotsService {
-
-  override func fetchCurrentSpot(callback: (ParsedSpot)->(), errorCallback: (NSError)->()) {
-
-    let spot = Spot.lucileSpot()
-    let parameters = [
-      "location": [
-        "type": "Point",
-        "coordinates": [spot.location!.coordinate.longitude, spot.location!.coordinate.latitude]
-      ],
-      "image_url": "http://example.com/image.jpg",
-      "user_id": 1,
-      "created_at": "2015-03-20T21:59:40.394Z",
-      "id": 1
-    ]
-
-    let parsedSpot = ParsedSpot(attributes: parameters)
-    callback(parsedSpot)
-  }
-
-  override func postNewSpot(spot: Spot, callback: (ParsedSpot)->(), errorCallback: (NSError)->()) {
-
-    let spot = Spot.griffithSpot()
-    let mockResponse = [
-      "image_url": "http://example.com/image.jpg",
-      "user_id": 2,
-      "created_at": "2015-03-20T21:59:40.394Z",
-      "id": 2
-    ]
-
-    let parsedSpot = ParsedSpot(attributes: mockResponse)
-    callback(parsedSpot)
-  }
-
-  override func postSpotGuess(guess: Guess, callback: (Bool)->(), errorCallback: (NSError)->()) {
-    callback(true)
-  }
-}
-
 class SpotTests: XCTestCase {
 
   override func setUp() {
